server/pyproject.toml
[tool.poetry]
name = "votes-server"
version = "0.1.0"
description = ""
authors = ["Rendy Arya Kemal <renrror@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Flask-JWT-Extended = "^4.4.0"
Authlib = "^1.0.1"
requests = "^2.27.1"
pydantic = "^1.9.0"
Flask-SQLAlchemy = "^2.5.1"
alembic = "^1.7.7"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.950"
sqlalchemy-stubs = "^0.4"
flask-sqlalchemy-stubs = {git = "https://github.com/qreserve/flask-sqlalchemy-stubs.git"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
